tests/wpt/tests/html/browsers/browsing-the-web/history-traversal/window-name-after-cross-origin-sub-frame-navigation.sub.html

<!DOCTYPE HTML>
<html>
<head>
    <!-- window.name should equal "test" after a cross-origin sub frame navigation. -->
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
</head>
<body>
    <script>
        t = async_test("Test that the window name is correct");
        window.addEventListener("message", t.step_func_done(function(e) {
            assert_equals(e.data, true);
        }));
    </script>
    <iframe src="support/window-name-navigation.sub.html?hostname={{domains[www1]}}&shouldhavename=true&sendmessage=true";
    </iframe>
</body>
</html>
